feat:add jwt auth middleware, verify /posts routes

+++ backend/routes/posts.mjs
@@ -1,38 +1,71 @@
 import express from "express";
 
 import db from "../db/db.mjs";
+import { authenticateToken } from "./users.mjs";
 import { ObjectId } from "mongodb";
+import { checkSchema } from "express-validator";
 
 const router = express.Router();
 
 const POSTS_COLLECTION = "posts";
 
+const postCreateUpdateSchema = {
+  title: {
+    errorMessage: "Empty post title is invalid",
+    notEmpty: true,
+  },
+  body: {
+    errorMessage: "Empty post body is invalid",
+    notEmpty: true,
+  },
+  author: {
+    errorMessage: "Empty post author is invalid",
+    notEmpty: true,
+  },
+};
+
+router.use(authenticateToken);
+
 // First 10 posts
 router.get("/", async (req, res) => {
   let collection = db.collection(POSTS_COLLECTION);
-  let postResults = await collection.find({}).limit(10).toArray();
+  let postResults = await collection
+    .find({ author: { $ne: req.user } })
+    .limit(10)
+    .toArray();
 
-  res.send(postResults).status(200);
+  res.json(postResults).status(200);
 });
 
 router.get("/all", async (req, res) => {
   let collection = db.collection(POSTS_COLLECTION);
-  let postResults = await collection.find({}).toArray();
+  let postResults = await collection
+    .find({ author: { $ne: req.user } })
+    .toArray();
 
-  res.send(postResults).status(200);
+  res.json(postResults).status(200);
 });
 
-router.post("/create", async (req, res) => {
-  let collection = db.collection(POSTS_COLLECTION);
+router.post(
+  "/create",
+  checkSchema(postCreateUpdateSchema),
+  async (req, res) => {
+    if (req.body.author != req.user) {
+      return res
+        .json("creating a post with another username is an invalid operation")
+        .status(400);
+    }
 
-  let newPost = req.body;
-  console.log(`Received posts/create with body: `);
-  console.log(req.body);
-  newPost.date = new Date();
+    let collection = db.collection(POSTS_COLLECTION);
+    let newPost = req.body;
+    console.log(`Received posts/create with body: `);
+    console.log(req.body);
+    newPost.date = new Date();
 
-  let result = await collection.insertOne(newPost);
-  res.send(result).status(204);
-});
+    let result = await collection.insertOne(newPost);
+    res.json(result).status(204);
+  },
+);
 
 router.get("/:id", async (req, res) => {
   const query = { _id: ObjectId.createFromHexString(req.params.id) };
@@ -40,25 +73,35 @@ router.get("/:id", async (req, res) => {
   let collection = db.collection(POSTS_COLLECTION);
   let result = await collection.findOne(query);
 
-  res.send(result).status(200);
+  res.json(result).status(200);
 });
 
-router.put("/:id", async (req, res) => {
-  const query = { _id: ObjectId.createFromHexString(req.params.id) };
+router.put("/:id", checkSchema(postCreateUpdateSchema), async (req, res) => {
+  if (req.body.author != req.user) {
+    return res
+      .json("updating a post with another username is an invalid operation")
+      .status(400);
+  }
 
+  const query = { _id: ObjectId.createFromHexString(req.params.id) };
   let collection = db.collection(POSTS_COLLECTION);
   let result = await collection.replaceOne(query, req.body);
 
-  res.send(result).status(200);
+  res.json(result).status(200);
 });
 
 router.delete("/:id", async (req, res) => {
-  const query = { _id: ObjectId.createFromHexString(req.params.id) };
+  if (req.body.author != req.user) {
+    return res
+      .json("deleting a post with another username is an invalid operation")
+      .status(400);
+  }
 
+  const query = { _id: ObjectId.createFromHexString(req.params.id) };
   let collection = db.collection(POSTS_COLLECTION);
   let result = await collection.deleteOne(query);
 
-  res.send(result).status(200);
+  res.json(result).status(200);
 });
 
 export default router;
